# docs/swagger.yaml
basePath: /
definitions:
...
  models.Point:
    properties:
      id:
        example: 1
        type: integer
      location:
        example: 69.164529, 35.138287
        type: string
      name:
        example: Teriberka
        type: string
    type: object
  models.User:
    properties:
      email:
        example: alex@example.com
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Alex
        type: string
      password:
        example: password
        type: string
    type: object
host: localhost:8081
info:
  contact: {}
  title: Travel Planner
  version: "1.0"
paths:
...
  /points:
    delete:
      description: Удаление точки по ID
      parameters:
      - description: ID точки
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Точка удалена
        "400":
          description: Неверный ID
          schema:
            type: string
        "404":
          description: Точка не найдена
          schema:
            type: string
      summary: Удалить точку
      tags:
      - Контрольные точки
    get:
      description: Получить все точки или конкретную точку по ID
      parameters:
      - description: ID точки
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ для одной точки
          schema:
            $ref: '#/definitions/models.Point'
        "400":
          description: Неверный ID
          schema:
            type: string
        "404":
          description: Точка не найдена
          schema:
            type: string
      summary: Получить точку/точки
      tags:
      - Контрольные точки
    patch:
      consumes:
      - application/json
      description: Частичное или полное обновление данных точки
      parameters:
      - description: ID точки
        in: query
        name: id
        required: true
        type: integer
      - description: Обновленные данные
        in: body
        name: point
        required: true
        schema:
          $ref: '#/definitions/models.Point'
      produces:
      - application/json
      responses:
        "200":
          description: Обновленная точка
          schema:
            $ref: '#/definitions/models.Point'
        "400":
          description: Неверные данные
          schema:
            type: string
        "404":
          description: Точка не найдена
          schema:
            type: string
        "409":
          description: Конфликт данных
          schema:
            type: string
      summary: Обновить точку
      tags:
      - Контрольные точки
    post:
      consumes:
      - application/json
      description: Добавление новой географической точки
      parameters:
      - description: Данные точки
        in: body
        name: point
        required: true
        schema:
          $ref: '#/definitions/models.Point'
      produces:
      - application/json
      responses:
        "201":
          description: Созданная точка
          schema:
            $ref: '#/definitions/models.Point'
        "400":
          description: Неверный формат данных
          schema:
            type: string
        "409":
          description: Конфликт данных
          schema:
            type: string
      summary: Создать новую точку
      tags:
      - Контрольные точки
  /users:
    delete:
      description: Удаление пользователя по ID
      parameters:
      - description: ID пользователя
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Пользователь удален
        "400":
          description: Некорректный ID
          schema:
            type: string
        "404":
          description: Пользователь не найден
          schema:
            type: string
      summary: Удалить пользователя
      tags:
      - Пользователи
    get:
      description: Возвращает конкретного пользователя по ID или всех пользователей,
        если ID не указан
      parameters:
      - description: ID пользователя
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Данные пользователя
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Некорректный ID
          schema:
            type: string
        "404":
          description: Пользователь не найден
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      summary: Получить пользователя/пользователей
      tags:
      - Пользователи
    patch:
      consumes:
      - application/json
      description: Частичное обновление данных пользователя. Пароль будет автоматически
        хеширован.
      parameters:
      - description: ID пользователя
        in: query
        name: id
        required: true
        type: integer
      - description: Обновляемые данные
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: Обновленные данные
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Некорректные данные
          schema:
            type: string
        "404":
          description: Пользователь не найден
          schema:
            type: string
        "409":
          description: Конфликт данных
          schema:
            type: string
      summary: Обновить пользователя
      tags:
      - Пользователи
    post:
      consumes:
      - application/json
      description: Регистрация нового пользователя в системе
      parameters:
      - description: Данные пользователя
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: Созданный пользователь
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Некорректные входные данные
          schema:
            type: string
        "409":
          description: Пользователь уже существует
          schema:
            type: string
        "500":
          description: Ошибка хеширования пароля
          schema:
            type: string
      summary: Создать пользователя
      tags:
      - Пользователи
swagger: "2.0"